Drop lwip-namespace.h renaming for VBoxNetNAT.  It is only necessary in VBoxDD to avoid conflicts with slirp.
#undef LWIP_PREFIX_BYTEORDER_FUNCS - its meaning was inverted in
lwip-1.4 and lwip-namespace.h has to redo the same redefinition dance
that lwip's def.h does when LWIP_PREFIX_BYTEORDER_FUNCS is undefined
(this is to be addressed for VBoxDD separately).

Debugging is more comfortable when names are not randomly renamed with
some of them being original short names and some redefined with lwip_
prefix.

Almost the same object code is generated, with slighly different
offsets into .rodata because of changed __func__.